Fixed enrollcommand test
