--- project/src/global.json
@@ -1,6 +1,0 @@
-{
-    "sdk": {
-      "version": "8.0.204"
-    }
-  }
-
